+++ ft_strncpy.c
@@ -6,7 +6,7 @@
 /*   By: bwan-nan <bwan-nan@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/11/07 17:58:30 by bwan-nan          #+#    #+#             */
-/*   Updated: 2018/11/07 17:58:32 by bwan-nan         ###   ########.fr       */
+/*   Updated: 2018/11/14 13:03:16 by bwan-nan         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,7 +15,7 @@
 
 char	*ft_strncpy(char *dest, const char *src, size_t n)
 {
-	int i;
+	size_t i;
 
 	i = 0;
 	while (i < n && src[i])
